fix: return type of mock UART available() should be size_t

+++ components/altherma_hub/mock_uart.h
@@ -18,7 +18,7 @@ namespace esphome
                 generate_response(data[2]);
             }
 
-            int available() override {
+            size_t available() override {
                 return rx_buffer_.size();
             }
